third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-insert-source-networkState-expected.txt

This is a testharness.js-based test.
FAIL NOT invoking resource selection by inserting <source> when networkState is not NETWORK_EMPTY assert_equals: networkState in second script expected 2 but got 3
Harness: the test ran to completion.
